requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
django-filter>=23.5
Pillow>=10.1.0
python-magic>=0.4.27
requests>=2.31.0
python-dotenv>=1.0.0
argparse>=1.4.0
pandas>=2.1.4
drf-nested-routers>=0.93.4
